Login Page
I have done the login page but came across a few challenges on the ".css" hence the new file. Firstly,  I had to create another div that would host the input fields, its the same with the one I used for the Home page but I jut couldn't get it to be at the center , hence the new one and somehow repeating code , I am still trying to get around that . Another thing you'll notice is that the size of the search area somehow increases as you move from one page to the other, I think  its because I have a certain height for the input area for the email and user name.